Allow more RAM for Display.Emulator.Mod
Move the display base to 3FE7F00H, which is sufficient for 64 MB of RAM.
If there is less RAM, in the JS emulator the RAM will "wrap around", so
it will still work.

As other emulators do not limit the virtual address space to the size of
the actual RAM, look for the magic value at bothe 0E7F00H and 3FE7F00H
and adjust the base accordingly.